flask/requirements.txt
blinker==1.8.2
certifi==2024.7.4
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
Faker==30.3.0
Flask==3.0.3
idna==3.7
itsdangerous==2.2.0
Jinja2==3.1.4
MarkupSafe==2.1.5
mysql-connector-python==9.0.0
pyTelegramBotAPI==4.21.0
python-dateutil==2.9.0.post0
requests==2.32.3
six==1.16.0
telebot==0.0.5
typing_extensions==4.12.2
urllib3==2.2.2
Werkzeug==3.0.3
